Add files via upload
Here is I would like to share the brief presentation of the thesis of Master's degree that investigated #housing #price of Sydney. Since 1976, #housing #pricing #storytelling is illustrated in the presentation. The majority of the challenge was part of "what macro factors affect #housingmarket" in the thesis. If you are interested in I will always be open to discussing with you.